refactor: move pipeline page helpers
