+++ tests/test_eval.py
@@ -11,7 +11,7 @@ from prml_vslam.eval.contracts import (
     EfficiencyEvaluationSelection,
 )
 from prml_vslam.eval.protocols import DenseCloudEvaluator, EfficiencyEvaluator, TrajectoryEvaluator
-from prml_vslam.eval.services import TrajectoryEvaluationService
+from prml_vslam.eval.services import TrajectoryEvaluationService, compute_trajectory_ape_preview
 from prml_vslam.utils import PathConfig
 
 
@@ -71,6 +71,18 @@ def test_trajectory_evaluation_service_satisfies_protocol(tmp_path: Path) -> Non
     assert isinstance(service, TrajectoryEvaluator)
 
 
+def test_compute_trajectory_ape_preview_returns_matched_error_series(tmp_path: Path) -> None:
+    reference_path = tmp_path / "reference.tum"
+    estimate_path = tmp_path / "estimate.tum"
+    _write_tum(reference_path, [(0.0, 0.0, 0.0, 0.0), (0.1, 1.0, 0.0, 0.0)])
+    _write_tum(estimate_path, [(0.0, 0.0, 0.0, 0.0), (0.1, 1.1, 0.1, 0.0)])
+
+    preview = compute_trajectory_ape_preview(reference_path=reference_path, estimate_path=estimate_path)
+
+    assert len(preview.error_series.values) == 2
+    assert preview.stats.rmse > 0.0
+
+
 def test_dense_cloud_protocol_accepts_structural_implementations(tmp_path: Path) -> None:
     service = FakeDenseCloudEvaluationService()
     selection = DenseCloudEvaluationSelection(
@@ -89,3 +101,11 @@ def test_efficiency_protocol_accepts_structural_implementations(tmp_path: Path) 
 
     assert isinstance(service, EfficiencyEvaluator)
     assert service.compute_efficiency_evaluation(selection=selection).metrics["latency_ms"] == 125.0
+
+
+def _write_tum(path: Path, rows: list[tuple[float, float, float, float]]) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(
+        "\n".join(f"{timestamp} {tx} {ty} {tz} 0 0 0 1" for timestamp, tx, ty, tz in rows),
+        encoding="utf-8",
+    )
